[FIX] Fixing begin/end calls in alignment module.
Was: Type retrieved with Iterator<, Rooted> and call to begin/end without rooted.

This leads to inconsistencies since the default iterator type is not rooted but determined by another metafunction.

Changed to: Explicitely give Rooted() in begin/end calls.

Now we can also use ModifiedString for reversing in align_split instead of in-place reverting.

git-svn-id: http://svn.seqan.de/seqan/trunk@14268 e6417c60-b987-48fd-844e-b20f0fcc1017