[DOC,FEATURE] Adding demo output to repository.
If there is a file FILE.cpp.stdout or FILE.cpp.stderr for a demo file FILE.cpp then
we will compare the contents of these files to the output of the program to standard
output or error besides checking the result code.

The .stdout and .stderr files can also be included from dox using `@include` to make
the comments in the documentation slimmer.

git-svn-id: http://svn.seqan.de/seqan/trunk@14420 e6417c60-b987-48fd-844e-b20f0fcc1017