[FIX,API] make const StringSets return const members
This patch defines the values in a const stringset to be const as well.
This behaviour is different to before, where you could have a const StringSet
of non-const Strings. The new behaviour is correct and intended according
weese and was required to make certain constructs like ReverseComplements
on members of ConcatDirect strings possible.

This patch was reviewed by weese. 



git-svn-id: http://svn.seqan.de/seqan/trunk@14447 e6417c60-b987-48fd-844e-b20f0fcc1017