[FEATURE,TEST] add translation module
This module includes code to do nucleaic acid translation to amino acid alphabet. The only things visible to the user are the translate function, an options enum and tags for the different translation tables. All currently used translation tables according to ncbi are included.
The implementation uses OpenMP if not disabled.

The code is fast :) Using OpenMP and ConcatDirect Sets for input and output it is possible to translate ~14million reads of ~100bp length in 2.3seconds on sequoia (excluding IO-time, including time for memory allocation of the target StringSet).




git-svn-id: http://svn.seqan.de/seqan/trunk@14448 e6417c60-b987-48fd-844e-b20f0fcc1017