[DOX,FIX] Adding @val tag to dox documentation.
Enum values now have to be documented using the @val tag instead of the
@var tag.

The problem with using @var for both variables and enum values is that
we cannot have variables with enum types.  Doxygen can resolve the
ambiguity by interpreting the C++ source code but we cannot.

This patch also updates the library's @var documentation entries to @val
when required.

Also updated the Dox API documentation to reflect this:

  http://trac.seqan.de/wiki/StyleGuide/DoxApiDocs