[FIX] Fixing mason2 on Windows
The fix looks quite esoteric and I think that the problem is also a bit esoteric.
I am not sure yet whether we did something wrong in SeqAn, or the MSVC optimizer
is a bit too harsh in is optimization.  In any case, it appears to be related to
using seqan::maxValue<int>() instead of seqan::MaxValue<int>::VALUE.  Note that
we do not have to replace all occurences but one is enough for each int and unsigned.