[FEATURE] Improved the runtime and quality of deep alignments by changing the threshold
For now there is no clear rule what is a deep alignment and how to choose the threshold. The changes have been made are based on experimental results for Balibase/Balifam benchmark containing different types of sequences. These parameters showed the best results in this benchmark, so we decided to keep them like it is. For not deep aligments we keep using the old parameter which also showed the best results.