[FIX] make mason2 work with changes to ModString
@holtgrewe ModifiedString does not contain a holder, but
only a pointer to the host therefore it can never be used
with temporaries. Unfortunately the old constructors
allowed this and will likely have worked in most situations,
but the host pointer was then always pointing at abandoned
stack.
The new constructors don't allow this, so we found the issue
and fixed in mason. Please review.