[FIX] Fixes #2417
In local alignment the comparison against 0 was fixed when
computing all traces. The rational behind this is that
whenever the score at a cell is not greater than 0 the trace
must be set to NONE to indicate the begin of the
alignment. Otherwise, a scoring matrix that evaluates to
score 0 for two residues might falsely add the substitution
to the alignment.