[FIX] gcc-14: bogus stringop-truncation
The idea of the warning is to warn when a null-terminated string is
copied without the null-terminator (e.g., a prefix). If this copied
string is then used, it is not null-terminated anymore, thus
portentially encountering UB.

However, this one is bogus because we actually do add a null-terminator
at pos - 1.
We circumvent this wrong diagnostic by using memcpy instead.