[FIX] heap-use-after-free
In the original line `appendValue(taskQueue, *taskGraph[0][0])`, the
std::shared_ptr was dereferenced and appended to the taskQueue.
This taskQueue was then used by the executor. On rare occassions
(depending on which thread does what and when), it could happen that the
executor tries to access the task* after it was destructed when all
std::shared_ptr instances were destructed.

By passing the task via shared_ptr instead of raw pointer, we can ensure
that any access by the executor is valid.